LayoutTests/media/modern-media-controls/aspect-ratio-button/aspect-ratio-button.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.css" type="text/css" media="screen">
<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/scheduler.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-node.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/layout-item.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-service.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/icon-button.js" type="text/javascript"></script>
<script src="../../../../Source/WebCore/Modules/modern-media-controls/controls/aspect-ratio-button.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>AspectRatioButton</code> class.");

iconService.directoryPath = "../../../../Source/WebCore/Modules/modern-media-controls/images";

const scaleToFillButton = new AspectRatioButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("scaleToFillButton.element.localName", "button");
shouldBeTrue('scaleToFillButton.element.classList.contains("icon")');
shouldBeTrue('scaleToFillButton.element.classList.contains("aspect-ratio")');
shouldBeEqualToString("scaleToFillButton.iconName", `${Icons.ScaleToFill}`);

const scaleToFitButton = new AspectRatioButton({ layoutTraits: LayoutTraits.macOS });
scaleToFitButton.scalesToFill = true;
shouldBeEqualToString("scaleToFitButton.iconName", `${Icons.ScaleToFit}`);

scheduler.frameDidFire = function()
{
    if ([scaleToFillButton, scaleToFitButton].some(button => button.width == 0))
        return;

    shouldBeTrue('scaleToFillButton.element.style.webkitMaskImage.includes("macOS/scale-to-fill@")');
    shouldBeTrue('scaleToFitButton.element.style.webkitMaskImage.includes("macOS/scale-to-fit@")');

    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
