LayoutTests/media/modern-media-controls/aspect-ratio-button/aspect-ratio-button-expected.txt

Testing the AspectRatioButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scaleToFillButton.element.localName is "button"
PASS scaleToFillButton.element.classList.contains("icon") is true
PASS scaleToFillButton.element.classList.contains("aspect-ratio") is true
PASS scaleToFillButton.iconName is "scale-to-fill"
PASS scaleToFitButton.iconName is "scale-to-fit"
PASS scaleToFillButton.element.style.webkitMaskImage.includes("macOS/scale-to-fill@") is true
PASS scaleToFitButton.element.style.webkitMaskImage.includes("macOS/scale-to-fit@") is true
PASS successfullyParsed is true

TEST COMPLETE
